backend/requirements.txt
Flask==2.1.2
PyPDF2==3.0.0
langchain==0.0.146
openai==0.27.0
Werkzeug==2.2.3
tabula-py==2.3.0
tiktoken==0.7.0